requirements.txt
streamlit>=1.32
streamlit-webrtc>=0.50.0
#opencv-python
av            # PyAV
#numpy
gdown         # 重みDL用
huggingface_hub[cli]   # モデル重みダウンロード用
ffmpeg-python          # 動画入出力用
diffusers
transformers
accelerate
# torch>=2.0.1